fix merge conflict refactored directory structure fix code style
