+++ app.js
@@ -2,10 +2,9 @@ import express from "express";
 import cors from "cors";
 import userRouter from "./routes/userRouter.js";
 
-import errorHandler from "./utils/errorHandler.js";
-import ErrorResponse from "./utils/ErrorResponse.js";
-
 import authRouter from "./routes/authRouter.js";
+import ErrorResponse from "./utils/errorHandlers/ErrorResponse.js";
+import errorHandler from "./utils/errorHandlers/errorHandler.js";
 
 const app = express();
 app.use(cors());
