--- utils/ErrorResponse.js
@@ -1,6 +1,0 @@
-export default class ErrorResponse extends Error {
-  constructor(message, statusCode) {
-    super(message);
-    this.statusCode = statusCode;
-  }
-}
