--- utils/asyncHandler.js
@@ -1,7 +1,0 @@
-export default function asyncHandler(controller) {
-  return (req, res, next) => {
-    controller(req, res, next).catch((error) => {
-      next(error);
-    });
-  };
-}
